% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{helpmds}
\alias{helpmds}
\title{Find an SDMX query code, or find out why yours does not work}
\usage{
helpmds(query = "", pattern = "", dim = NULL, verbose = TRUE)
}
\arguments{
\item{query}{a character query string with a RestFul SDMX query: either empty string, or a provider such as ('ECB' or 'LTNSI'), or provider.dataflow (such as 'ECB/EXR' ) or a full query  (such as ECB/EXR/A.GBP+CHF.EUR.SP00.A). below for details}

\item{pattern}{a character string to search for in the result that \code{query} provides}

\item{dim}{a character string denoting a certain dimension of the dataflow indicated by query (or the integer number of that dimension),}

\item{verbose}{if TRUE, the function will chat about servers requests and warn when it might take some time}
}
\value{
the function prints explanations and invisibly returns a list or data.frame in addition
}
\description{
Find an SDMX query code, or find out why yours does not work
}
\details{
This function works with the metadata of the SDMX providers, which can be slow. Note that in many cases \code{\link[MD3]{helpNomics}} might be faster to use and yield the same result.
}
\section{SDMX query}{

A restful SDMX query combines Provider, Dataflow and Dimension selectors. Take the example 'ECB/EXR/A.GBP+CHF.EUR.SP00.A'
\itemize{
  \item 'ECB' designates the provider ECB (see helpmds() for a list of providers)
  \item 'EXR' designates the ECB's dataflow 'EXR' about exchange rates (see helpmds('ECB') for a list of ECB dataflows)
  \item A.GBP+CHF.EUR.SP00.A is the dimension selector: See helpmds('ECB/EXR') for what dimensions exist in ECB/EXR)
}
Here,
\itemize{
\item 'A' denotes annual frequency from the FREQ dimension
\item 'GBP+CHF' denote British pound and Swiss frank from the currency dimension
\item 'EUR' denotes the currency denominator
\item 'SP00' denotes spot rates
\item 'A' denotes average over periods (here annual averages)
}
Omit dimension codes to load all elements of a dimension, i.e. ECB/EXR/A..EUR.SP00.A to download exchange ratesw of the euro vis-a-vis all currencies
Note that a query is normally taken as Provider/Dataflow/Selectors. The combination with dots Provider.Dataflow.Selectors is not tolerated here.
Note that dataflow codes may be case-sensitive
}

\section{Finding query codes}{


To find query codes, the easiest is to go to \url{http://db.nomics.world} and search for codes there
Alternatively, you can use \code{\link{helpNomics}}

\itemize{
 \item \code{helpmds()} returns a vector with available providers
 \item \code{helpmds("PROVIDER")} returns a vector with the dataflows available for the provider
 \item \code{helpmds("PROVIDER", pattern="PATTERN")} looks for a dataflow whose name or code matches the string pattern
 \item \code{helpNomics("PROVIDER/DATAFLOW")} prints the structure of the dataflow dimensions, as well as an example query, and invisible returns the structure of code names within the dataflow
 \item \code{helpNomics("PROVIDER/DATAFLOW/SELECTORS")} invisibly returns the query result. If not successful, it provides clues whwere the query might be wrong
 \item \code{helpNomics("PROVIDER/DATAFLOW", dim="NumberOrName")} prints the code names and labels of the selected dimension, and invidibly returns a data.frame with those codes and labels
 \item \code{helpNomics("PROVIDER/DATAFLOW", dim="NumberOrName", pattern="PATTERN")} searches the codes and labels of that dimension for a certain pattern
}
}

\examples{
#lets find house price data from the Bank of International Settlements
helpmds() #BIS is a provider that is available
helpmds(pattern='bis.org') #indeed

helpmds('BIS') #list all available dataflows

helpmds('BIS',pat='property') #so the dataflow we need is called 'WS_SPP'

helpmds('BIS/WS_SPP') # so there are 4 dimensions (apart form time): FREQ, REF_AREA, VALUE, UNIT_MEASURE

#lets check what is in ALUE
helpmds('BIS/WS_SPP',dim='VALUE') #only two dimensions in VALUE
oo=helpmds('BIS/WS_SPP',dim='REF_AREA') #but lots of countries. I guess not all of them will be filled
oo # show all countries in there

helpmds('BIS/WS_SPP',dim='FREQ') #all kinds of frequencies too that are permissible. But not all of them will have data.

foo=mds('BIS/WS_SPP/Q.FR.N.', labels=TRUE) #lets look for nominal quarterly house prices from France
foo #hmm, what does 628 mean?
helpmds('BIS/WS_SPP/Q.FR.N.') #one way to find out

dimcodes(foo)[[1]] #another way - using dimcodes

foo=helpmds('BIS/WS_SPP/Q.FR.N.') # another way to find out
#note that foo now contains the dimcodes for that code

helpmds('BIS/WS_SPP/Q.FR.N.', dim='UNIT_MEASURE') # another way to find out

helpmds('BIS/WS_SPP',dim=4,pattern='628' ) # another way to find out


#find out why another formulation does not work:
helpmds('BIS/WS_SPP/Q.FRA.N.')

}
\seealso{
\code{\link{mds}} to load data,  \code{\link[MD3]{helpNomics}} for a similar function
}
