% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useexternally.R
\name{loadExternally}
\alias{loadExternally}
\title{Function and script to make mds callable form otehr applications, such as Stata}
\usage{
loadExternally(
  ...,
  outpath = "C:/Users/Public/statautils/queryresult.dta",
  ascountrypanel = TRUE
)
}
\arguments{
\item{\dots}{either a character call to mds or a related function (such as ECB/EXR/A.JPY.EUR.SP00.A), or a list containing that call, aswell as outpath and ascountrypanel}

\item{outpath}{Character. the dta or csv file to be created be the function (default C:/Users/Public/statautils/queryresult.dta). If empty, nullor empty string, then loadExternally simply returns the object on the R console}

\item{ascountrypanel}{logical (Default TRUE). If TRUE, then this attempts to reshape the result into a panel form with "variables" being the columns, and caountries and TIME being identified in the first column. If FALSE, then teh result is fully stacked, and contains observation attributes.}
}
\value{
the file path to outpath, or a data.table (if outpath=='')
}
\description{
This function is meant to be called from other applications via RScript and arguments. If you don't know R look below under the section Stata Users resp Non-Stata Users
}
\details{
if left empty, then this function creates the script loadMD.R to be called with R script.
The default path for that script is C:/Users/Public/statautils/loadMD.R but than be changed using outpath.
In the latter case like loadExternally(outpath='c:/balbal.R') can also be used in non-interactive mode
}
\section{For Stata Users that dont know R}{


 1) make sure you have internet connection from R. Run e.g. \code{readLines('https://example.com',n = 1)} to check that
 2) make sure you have readstata13 installed. if \code{library(readstata13)} does not work, call \code{install.packages('readstata13')}
 3) run\code{loadExternally()}. Press Enter, and then type \code{s} and Enter

 Now from Stata, type e.g.

 \code{shell "C:\\Users\\Public\\statautils\\loadmd" ECB/EXR/A.JPY+PLN.EUR.SP00.A}
 \code{use "C:\\Users\\Public\\statautils\\queryresult.dta", clear}

 or

 \code{shell "C:\\Users\\Public\\statautils\\loadmd" AMECO/C/BE+AT.1_0_0_0_UVGD 0}
 \code{use "C:\\Users\\Public\\statautils\\queryresult.dta", clear}

 Pro-tip: IF you want all country codes to look the same (e.g. all are UK rather than GB or GBR), then uncomment the line with \code{defaultcountrycode} in C:/Users/Public/statautils/loadMD.R
}

\section{For non-Stata Users that dont know R}{


 1) make sure you have internet connection from R. Run e.g. \code{readLines('https://example.com',n = 1)} to check that
 2) make sure you have readstata13 installed. if \code{library(readstata13)} does not work, call \code{install.packages('readstata13')}
 3) run\code{loadExternally()}. Press Enter, and then type \code{a} and Enter

 Now from Shell, type e.g.

 \code{C:\\Users\\Public\\statautils\\loadmd ECB/EXR/A.JPY+PLN.EUR.SP00.A}

 or

 \code{C:\\Users\\Public\\statautils\\loadmd AMECO/C/BE+AT.1_0_0_0_UVGD 1 c:/Users/Public/test.csv}
 to create a CSV file called \code{test.csv}

 Pro-tip: IF you want all country codes to look the same (e.g. all are UK rather than GB or GBR), then uncomment the line with \code{defaultcountrycode} in C:/Users/Public/statautils/loadMD.R
}

\examples{
#creates the script loadMD.R:
\dontrun{loadExternally()}

loadExternally('nomics:ECB/EXR/A.JPY+USD.EUR.SP00.E+A',outpath='')
loadExternally('ECB/EXR/A.JPY+PLN+HUF.EUR.SP00.A+E',outpath='')
loadExternally('ECB/EXR/A.JPY+PLN+HUF.EUR.SP00.A+E',outpath='', ascountrypanel=FALSE)

loadExternally('ESTAT/prc_hpi_a/A.TOTAL..BE')
loadExternally(list('ESTAT/prc_hpi_a/A.TOTAL..BE','',FALSE))
loadExternally('ECB/QSA/Q.N..W0.S1M.S1.N.L.LE.F3+F4.T._Z.XDC._T.S.V.N._T')





}
\seealso{
\code{\link{helpmds}} if something does not work. e.g. \code{helpmds('ESTAT/prc_hpi_a/A.TOTAL..AT.')} to find out why that code does not work
 \code{\link{mds}} for loading from major data sources.
}
