% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecameco.R
\name{mdAmeco}
\alias{mdAmeco}
\title{Load data from publicly available Ameco, and is past vintages}
\usage{
mdAmeco(
  code = "",
  year = 0,
  release = 0,
  as = c("md3", "array", "numeric", "data.table", "zoo", "2d", "1d"),
  drop = TRUE,
  ccode = defaultcountrycode(),
  startPeriod = NULL,
  endPeriod = NULL,
  inclaggreg = FALSE
)
}
\arguments{
\item{code}{a RestFul code combination respecting Ameco dimensions GEO.x.x.x.x.INDICATOR, e.g. \code{CAN.1.0.99+0.0.UVGD}}

\item{year}{year of the release. Default 0 means current/latest release}

\item{release}{character singleton 'autumn' or 'spring'. Ignored if year is 0 and thus points to current release.}

\item{as}{how to output the result: \code{md3}: as md3 object with full metadata, \code{2d}: as a data.table with periods as column names, \code{1d} resp. \code{data.table}: as fully stacked data.table, \code{array}: as multi-dim array, \code{zoo}: as \code{\link[zoo]{zooreg}} time series object.}

\item{drop}{if TRUE, drop any singleton dimensions (see also drop.md3)}

\item{ccode}{If not \code{ccode==NULL}, then the function attempts to convert to ccode destination such as 'iso2c', 'EC', 'iso3c'. See \code{\link[MDcountrycode]{ccode}} for permissible values.  \code{\link[MDcountrycode]{defaultcountrycode}} for defining this value as a session-wide option.}

\item{startPeriod}{startyear (1960 if empty)}

\item{endPeriod}{end year (latest available if left empty)}

\item{inclaggreg}{default FALSE. In that case, if all countries are requested, the function returns individual counties as well as the latest EU and EA aggregate. If TRUE, it also returns other aggregates from Ameco, such as EA12.}
}
\value{
an md3 object or other as specified by \code{as}
}
\description{
Load data from publicly available Ameco, and is past vintages
}
\details{
This function works by caching Ameco vintages on the local drive
}
\examples{

mdAmeco('CAN.1.0.99+0.0.UVGD') #Canadian GDP in EUR and CAD

foo=mdAmeco('.1.0.0.0.ZCPIH') #consumer price index in all countries
\dontrun{foo['.y2019:y2023']}

foo=mdAmeco()
\dontrun{foo['DEU+ESP.1_0_310_0.UBCA.']} #German and Spanish current account as percent of GDP

\dontrun{growth(mdAmeco('EST+CZE.1.0.0.0.ZCPIH',startPeriod=2020))} #inflation rates in Estonia and Czechia from 2021



}
\seealso{
\code{\link{mdStat}} for loading from major data sources
}
