% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mdSdmx}
\alias{mdSdmx}
\title{Get data from an SDMX source}
\usage{
mdSdmx(
  code,
  startPeriod = "",
  endPeriod = "",
  drop = TRUE,
  labels = FALSE,
  as = c("md3", "array", "numeric", "data.table", "zoo", "2d", "1d", "pdata.frame",
    "data.frame"),
  ccode = getOption("defaultcountrycode", NULL),
  verbose = FALSE
)
}
\arguments{
\item{code}{a character query string with a RestFul SDMX query (such as ECB/EXR/A.GBP+CHF.EUR.SP00.A). See \code{\link{mds}} for an example}

\item{startPeriod}{optional, character, integer, timo or date-like class. default empty string means to fetch data including  the first available period}

\item{endPeriod}{placeholder fo optional, character, integer, timo or date-like class. default empty string means to fetch data including  the last available observation}

\item{drop}{if TRUE, drop any singleton dimensions (see also drop.md3)}

\item{labels}{add the descriptions for each dimension element, e.g. "Belgium" for "BEL". You can check those with function \code{\link[MD3]{dimcodes}}}

\item{as}{how to output the result: \code{md3}: as md3 object with full metadata, \code{2d}: as a data.table with periods as column names, \code{1d} resp. \code{data.table}: as fully stacked data.table, \code{array}: as multi-dim array, \code{zoo}: as \code{\link[zoo]{zooreg}} time series object.}

\item{ccode}{ccode If not \code{ccode==NULL}, then the function attempts to convert to ccode destination such as 'iso2c', 'EC', 'iso3c'. See \code{\link[MDcountrycode]{ccode}} for permissible values.  \code{\link[MDcountrycode]{defaultcountrycode}} for defining this value as a session-wide option.}

\item{verbose}{if TRUE, this function chatters about what it is doing, notably contacting servers}
}
\value{
depending on \code{as}, an \code{md3}, \code{array}, \code{numeric}, \code{data.table}, \code{zoo}, \code{2d} dat.table, \code{1d} stacked data.table, or data.frame    containing the requested data
}
\description{
Get data from an SDMX source
}
\details{
\code{mdSdmx} is a function for loading from SDMX data sources. \code{\link{mds}} encompasses that one, and will be broadened to other data sources at a later stage.
}
\examples{

foo=mdSdmx("ESTAT/prc_hpi_q/Q..I15_Q.")
foo[AT:BG.TOTAL.y2020:y]

mdSdmx("BBK/BBFBOPV/M.N.DE.W1.S1.S1.T.B.G+S+CA._Z._Z._Z.EUR._T._X.N.ALL")
mdSdmx('IMF/FSI/A.FI+DK+AT.FSASDLD_EUR+FSANL_PT')

mdSdmx("BIS/WS_EER/M.N.B.AT+FR+BE",startPeriod=2021)
#is the same as
mds("BIS/WS_EER/M.N.B.AT+FR+BE",startPeriod=2021)

#but mds can also do providers that helpmds() marks with an asterisk, such as AMECO
}
\seealso{
\code{\link{helpmds}}, \code{\link{DTstat}}, \code{\link[MD3]{Nomics}}, \code{\link[MDcountrycode]{defaultcountrycode}}, \code{\link{mdAmeco}}, \code{\link{mdWEO}}
}
