% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDstats.R
\name{MDstats}
\alias{MDstats}
\title{MDstats: A package providing for multi-dimensional data manipulation}
\description{
MDstats: A package providing for multi-dimensional data manipulation
}
\section{MDstats functionality}{

Blabla
}

\section{Further work}{



This packge is still under development; there are certainly some bugs lurking around

Conversion to and from \code{data.table}, tibble, and \code{pdata.frame} is to be added
}

\section{Author}{

Stefan Zeugner ECFIN.B1
}

