% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imfweo.R
\name{mdWEO}
\alias{mdWEO}
\title{Load data from IMF World Economic Outlook, and its past vintages}
\usage{
mdWEO(
  code = NULL,
  year = 0,
  release = 0,
  as = c("md3", "array", "numeric", "data.table", "zoo", "2d", "1d"),
  drop = TRUE,
  ccode = NULL,
  startPeriod,
  endPeriod
)
}
\arguments{
\item{code}{a RestFul code combination respecting IMF WEO dimensions GEO.INDICATOR, e.g. \code{CAN+AUS.NID_NGDP}. If set to NULL.empty vector then the cfunction will retuner everything - the entire WEO}

\item{year}{year of the release. Default 0 means current/latest release}

\item{release}{numeric singleton 1 (April release) or 2 (October release). Ignored if year is 0 and thus points to current release.}

\item{as}{how to output the result: \code{md3}: as md3 object with full metadata, \code{2d}: as a data.table with periods as column names, \code{1d} resp. \code{data.table}: as fully stacked data.table, \code{array}: as multi-dim array, \code{zoo}: as \code{\link[zoo]{zooreg}} time series object.}

\item{drop}{if TRUE, drop any singleton dimensions (see also drop.md3)}

\item{ccode}{If not \code{ccode==NULL}, then the function attempts to convert to ccode destination such as 'iso2c', 'EC', 'iso3c'. See \code{\link[MDcountrycode]{ccode}} for permissible values.  \code{\link[MDcountrycode]{defaultcountrycode}} for defining this value as a session-wide option.}

\item{startPeriod}{startyear (1960 if empty)}

\item{endPeriod}{end year (latest available if left empty)}
}
\value{
an md3 object or other as specified by \code{as}
}
\description{
Load data from IMF World Economic Outlook, and its past vintages
}
\details{
This function works by caching IMF WEO vintages on the local drive
}
\examples{

mdWEO('CA+US+MX.LP') #Population of Canada, US and Mexico
mdWEO('CHN+JPN+KOR.LP') #also works with three-letter codes


foo=mdWEO(ccode='iso2m')
foo[DE+ES.NGDP.] #German and Spanish  GDP


}
\seealso{
\code{\link{mdStat}} for loading from major data sources
}
