% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{providertable}
\alias{providertable}
\title{Table with basic information about SMDX providers}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{http://example.com}{example},#'
}
\usage{
data(providertable)
}
\description{
to be expanded later
}
\examples{
data(providertable)
}
\keyword{datasets}
