% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useexternally.R
\name{loadExternally}
\alias{loadExternally}
\title{Function and script to make mds callable form other applications, such as Stata}
\usage{
loadExternally(
  ...,
  outpath = "C:/Users/Public/statautils/queryresult.dta",
  ascountrypanel = TRUE
)
}
\arguments{
\item{\dots}{either a character call (such as "ECB/EXR/A.JPY.EUR.SP00.A") to mds or a related function , or a list containing that call, as well as outpath and ascountrypanel}

\item{outpath}{Character. the dta or csv file to be created be the function (default C:/Users/Public/statautils/queryresult.dta). If empty, NULL, or empty string, then loadExternally simply returns the object on the R console}

\item{ascountrypanel}{logical (Default TRUE). If TRUE, then this attempts to reshape the result into a panel form with "variables" being the columns, and countries and TIME being identified in the first column. If \code{FALSE}, then the result is fully stacked, and contains observation attributes.}
}
\value{
the file path to outpath, or a data.table (if outpath=='')
}
\description{
This function is meant to be called from other applications via RScript and arguments. If you don't know R, look below under the section Stata Users resp Non-Stata Users
}
\details{
if left empty, then this function creates the script loadMD.R to be called with R script.
The default path for that script is C:/Users/Public/statautils/loadMD.R but than be changed using outpath.
(In the latter case like \code{loadExternally(outpath='c:/balbal.R')} can also be used in non-interactive mode)

this also creates the wrapper file loadMD.cmd which is easier to handle than using loadMD.R directly
You can call it by \code{C:/Users/Public/statutils/loadMD Arg1 Arg2 Arg3}. The first Argument \code{Arg1} is mandatory, whereas  \code{Arg2}  and  \code{Arg3} are optional

loadMD accepts the following arguments
\itemize{
\item{Arg1}{ (mandatory): code to be loaded via mds, like ESTAT/prc_hpi_a/A.TOTAL..BE. Can be prefixed with \code{nomics:}, \code{ameco:} or \code{imfweo}: to access specific functions other than \code{mds}}
\item{Arg2}{ (optional): 1 or 0.  Setting this to 1 (default) makes the final file a country panel, with geo and time dimensions in the first column
  and all other dimensions forming separate variables. Or 0:make the result fully stacked, with n identifier columns and one \code{obs_value} column}
\item{Arg3}{  (optional): the filepath where to save the result. Must be either .csv or .dta (Stata) file. Default is the \code{defaultoutpath} variable defined in loadMD.R}
}

 Pro-tip: IF you want all country codes to look the same (e.g. all are UK rather than GB or GBR), then uncomment the line with \code{defaultcountrycode} in C:/Users/Public/statautils/loadMD.R
 check \code{\link{defaultcountrycode}} for help on that.
}
\section{For Stata Users that dont know R}{


 \itemize{
  \item{1)}{ make sure you have internet connection from R. Run e.g. \code{readLines('https://example.com',n = 1)} to check that}
  \item{2)}{ make sure you have readstata13 installed. if \code{library(readstata13)} does not work, call \code{install.packages('readstata13')}}
  \item{3)}{ run \code{loadExternally()}. Press Enter, and then type \code{s} and Enter}
 }

 Now from Stata, type e.g.

 \code{shell "C:\\Users\\Public\\statautils\\loadmd" ECB/EXR/A.JPY+PLN.EUR.SP00.A}

 \code{use "C:\\Users\\Public\\statautils\\queryresult.dta", clear}

 or use the second argument to load this  in 'fully stacked' form rather than as a country panel:

 \code{shell "C:\\Users\\Public\\statautils\\loadmd" AMECO/C/BE+AT.1_0_0_0_UVGD 0}

 \code{use "C:\\Users\\Public\\statautils\\queryresult.dta", clear}

See section 'Details' above for explaining those arguments. See \code{\link{helpmds}} for understanding the structure of dataset codes and dimensions
}

\section{For non-Stata Users that dont know R}{

 \itemize{
  \item{1)}{ make sure you have internet connection from R. Run e.g. \code{readLines('https://example.com',n = 1)} to check that}
  \item{2)}{ run\code{loadExternally()}. Press Enter, and then type \code{a} and Enter}
 }
 Now from DOS/Shell/Console (Windows+R, then CMD), type e.g.

 \code{C:\\Users\\Public\\statautils\\loadmd ECB/EXR/A.JPY+PLN.EUR.SP00.A}

 or

 \code{C:\\Users\\Public\\statautils\\loadmd AMECO/C/BE+AT.1_0_0_0_UVGD 1 c:/Users/Public/test.csv}
 to create a CSV file called \code{test.csv}

 See section 'Details' above for explaining those arguments. See \code{\link{helpmds}} for understanding the structure of dataset codes and dimensions
}

\examples{
#creates the script loadMD.R:
\dontrun{loadExternally()}

loadExternally('nomics:ECB/EXR/A.JPY+USD.EUR.SP00.E+A',outpath='')
loadExternally('ECB/EXR/A.JPY+PLN+HUF.EUR.SP00.A+E',outpath='')
loadExternally('ECB/EXR/A.JPY+PLN+HUF.EUR.SP00.A+E',outpath='', ascountrypanel=FALSE)

loadExternally('ESTAT/prc_hpi_a/A.TOTAL..BE')
loadExternally(list('ESTAT/prc_hpi_a/A.TOTAL..BE','',FALSE))
loadExternally('ECB/QSA/Q.N..W0.S1M.S1.N.L.LE.F3+F4.T._Z.XDC._T.S.V.N._T')





}
\seealso{
\code{\link{helpmds}} if something does not work. e.g. \code{helpmds('ESTAT/prc_hpi_a/A.TOTAL..AT.')} to find out why that code does not work
 \code{\link{mds}} for loading from major data sources.
}
