% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDstats.R
\name{MDstats}
\alias{MDstats}
\title{MDstats: A package providing for loading rom SDMX and other macro data sources for multi-dimensional data manipulation}
\description{
MDstats: A package providing for loading rom SDMX and other macro data sources for multi-dimensional data manipulation
}
\section{MDstats functionality}{

Use function \code{mds} and function \code{helpmds} to access all
}

\section{Further work}{



This package is still under development; there are certainly some bugs lurking around
Build date: 2024-01-09


Conversion to and from \code{data.table}, tibble, and \code{pdata.frame} is to be added
}

\section{Author}{

Stefan Zeugner ECFIN.B1
}

